requirements.txt
# Robot Portrait Drawing System Dependencies

# Environment
python-dotenv>=1.0.0

# Robot control
pymycobot>=3.0.0
pyserial>=3.5

# Computer vision
opencv-python>=4.5.0
numpy>=1.21.0

# AI integration
openai>=1.0.0

# Path planning (optional, for advanced optimization)
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Image processing
Pillow>=9.0.0

# Web visualization
flask>=2.0.0
flask-socketio>=5.0.0
